+++ src/game_sp/DDScrollBuffer.h
@@ -8,8 +8,8 @@
 #include "global.h"
 
 
-extern long mScrollX, mScrollY;
-extern long mScrollX_last, mScrollY_last;
+extern int mScrollX, mScrollY;
+extern int mScrollX_last, mScrollY_last;
 
 #if 1
 extern boolean redraw[2 + MAX_PLAYFIELD_WIDTH + 2][2 + MAX_PLAYFIELD_HEIGHT + 2];
@@ -26,6 +26,6 @@ extern void RestorePlayfield();
 
 extern void DDScrollBuffer_ScrollTo(int X, int Y);
 extern void DDScrollBuffer_ScrollTowards(int X, int Y, double Step);
-extern void DDScrollBuffer_SoftScrollTo(int X, int Y, long TimeMS, int FPS);
+extern void DDScrollBuffer_SoftScrollTo(int X, int Y, int TimeMS, int FPS);
 
 #endif /* DDSCROLLBUFFER_H */
